+++ R/members.R
@@ -12,7 +12,7 @@ setMethod("[[<-", c("MemberCatalog", "ANY", "missing", "ANY"), .backstopUpdate)
 setMethod("[[<-", c("MemberCatalog", "character", "missing", "NULL"),
     function (x, i, j, value) {
         ## Remove the specified user from the catalog
-        payload <- sapply(i, function (z) NULL, simplify=FALSE)
+        payload <- sapply(i, null, simplify=FALSE)
         crPATCH(self(x), body=toJSON(payload))
         return(refresh(x))
     })
@@ -29,13 +29,27 @@ setMethod("members<-", c("CrunchTeam", "MemberCatalog"), function (x, value) {
 ##' @rdname teams
 ##' @export
 setMethod("members<-", c("CrunchTeam", "character"), function (x, value) {
-    payload <- sapply(value,
-        function (z) emptyObject(),
-        simplify=FALSE)
+    payload <- sapply(value, emptyObject, simplify=FALSE)
     crPATCH(self(members(x)), body=toJSON(payload))
     return(refresh(x))
 })
 
-##' @rdname describe-catalog
-##' @export
-setMethod("emails", "ShojiCatalog", function (x) getIndexSlot(x, "email"))
+setMethod("is.editor", "MemberCatalog", function (x) {
+    ## N.B.: this is for projects; teams don't work this way.
+    vapply(index(x), function (a) {
+            isTRUE(a[["permissions"]][["edit"]])
+        }, logical(1), USE.NAMES=FALSE)
+})
+
+setMethod("is.editor<-", c("MemberCatalog", "logical"), function (x, value) {
+    stopifnot(length(x) == length(value))
+    changed <- is.editor(x) != value
+    if (any(changed)) {
+        payload <- structure(lapply(value[changed], {
+            function (v) list(permissions=list(edit=v))
+        }), .Names=urls(x)[changed])
+        crPATCH(self(x), body=toJSON(payload))
+        x <- refresh(x)
+    }
+    return(x)
+})